python/pyproject.toml
[project]
name = "eeva"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "anyio>=4.9.0",
    "fastapi>=0.115.12",
    "hydra-core>=1.3.2",
    "langchain>=0.3.24",
    "langchain-anthropic>=0.3.18",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.14",
    "langgraph>=0.6.7",
    "pydantic>=2.11.3",
    "regex>=2024.11.6",
    "sse-starlette>=2.3.6",
    "tabulate>=0.9.0",
    "types-aiofiles>=24.1.0.20250822",
    "types-regex>=2025.9.18.20250921",
    "types-tabulate>=0.9.0.20241207",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "requests>=2.32.3",
    "ruff>=0.11.6",
    "types-requests>=2.32.0.20250515",
    "supabase>=2.16.0",
    "google-cloud-logging>=3.12.1",
    "matplotlib>=3.10.5",
    "ipykernel>=6.29.5",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.2.250827",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint]
ignore = ["D203", "D213", "F722"]
select = ["I", "E", "F", "W", "N", "B"]
isort.known-first-party = ["eeva"]
